more double promote fixes

+++ rtengine/hilite_recon.cc
@@ -323,7 +323,7 @@ void RawImageSource::HLRecovery_inpaint(float** red, float** green, float** blue
 
     if (settings->verbose) {
         for (int c = 0; c < 3; ++c) {
-            printf("chmax[%d] : %f\tclmax[%d] : %f\tratio[%d] : %f\n", c, chmax[c], c, clmax[c], c, chmax[c] / clmax[c]);
+            printf("chmax[%d] : %f\tclmax[%d] : %f\tratio[%d] : %f\n", c, static_cast<double>(chmax[c]), c, static_cast<double>(clmax[c]), c, static_cast<double>(chmax[c] / clmax[c]));
         }
     }
 
@@ -366,7 +366,7 @@ void RawImageSource::HLRecovery_inpaint(float** red, float** green, float** blue
 
     if (settings->verbose) {
         for (int c = 0; c < 3; ++c) {
-            printf("correction factor[%d] : %f\n", c, factor[c]);
+            printf("correction factor[%d] : %f\n", c, static_cast<double>(factor[c]));
         }
     }
